Add workflow instance model and document title
