+++ documents/models.py
@@ -28,6 +28,7 @@ class Document(TimestampedModel):
         (STATUS_COMPLETED, "abgeschlossen"),
     ]
 
+    title = models.CharField(max_length=255)
     file = models.FileField(upload_to="documents/")
     status = models.CharField(
         max_length=20, choices=STATUS_CHOICES, default=STATUS_UPLOADED
