+++ projects/admin.py
@@ -1,6 +1,6 @@
 from django.contrib import admin
 
-from .models import Project, WorkflowInstance
+from .models import Project
 
 
 @admin.register(Project)
@@ -16,21 +16,3 @@ class ProjectAdmin(admin.ModelAdmin):
         "updated_at",
     )
     list_filter = ("organization", "status")
-
-
-@admin.register(WorkflowInstance)
-class WorkflowInstanceAdmin(admin.ModelAdmin):
-    """Admin configuration for :class:`WorkflowInstance`."""
-
-    list_display = (
-        "project",
-        "organization",
-        "template",
-        "created_at",
-        "updated_at",
-    )
-    list_filter = ("project__organization", "template")
-
-    @staticmethod
-    def organization(obj):
-        return obj.project.organization
