+++ workflows/admin.py
@@ -1,6 +1,6 @@
 from django.contrib import admin
 
-from .models import WorkflowTemplate, WorkflowStep
+from .models import WorkflowTemplate, WorkflowStep, WorkflowInstance
 
 
 @admin.register(WorkflowTemplate)
@@ -16,3 +16,21 @@ class WorkflowStepAdmin(admin.ModelAdmin):
 
     list_display = ("template", "order", "created_at", "updated_at")
     list_filter = ("template",)
+
+
+@admin.register(WorkflowInstance)
+class WorkflowInstanceAdmin(admin.ModelAdmin):
+    """Admin configuration for :class:`WorkflowInstance`."""
+
+    list_display = (
+        "project",
+        "organization",
+        "status",
+        "created_at",
+        "updated_at",
+    )
+    list_filter = ("project__organization", "status")
+
+    @staticmethod
+    def organization(obj):
+        return obj.project.organization
