+++ workflows/models.py
@@ -1,6 +1,8 @@
 from django.db import models
 
 from common.models import TimestampedModel
+from projects.models import Project
+from organizations.query import OrganizationManager
 
 
 class WorkflowTemplate(TimestampedModel):
@@ -26,3 +28,28 @@ class WorkflowStep(TimestampedModel):
 
     def __str__(self) -> str:  # pragma: no cover - simple representation
         return f"{self.template} - step {self.order}"
+
+
+class WorkflowInstance(TimestampedModel):
+    """Tracks the workflow status for a :class:`Project`."""
+
+    STATUS_DRAFT = "draft"
+    STATUS_REVIEW = "review"
+    STATUS_FINAL = "final"
+
+    STATUS_CHOICES = [
+        (STATUS_DRAFT, "Draft"),
+        (STATUS_REVIEW, "Review"),
+        (STATUS_FINAL, "Final"),
+    ]
+
+    project = models.ForeignKey(
+        Project, related_name="workflows", on_delete=models.CASCADE
+    )
+    status = models.CharField(
+        max_length=10, choices=STATUS_CHOICES, default=STATUS_DRAFT
+    )
+    objects = OrganizationManager(organization_field="project__organization")
+
+    def __str__(self) -> str:  # pragma: no cover - simple representation
+        return f"Workflow {self.get_status_display()} for {self.project}"
