+++ workflows/tests/test_models.py
@@ -1,6 +1,10 @@
 import pytest
 
-from .factories import WorkflowStepFactory, WorkflowTemplateFactory
+from .factories import (
+    WorkflowInstanceFactory,
+    WorkflowStepFactory,
+    WorkflowTemplateFactory,
+)
 
 
 @pytest.mark.django_db
@@ -13,3 +17,10 @@ def test_workflow_template_factory():
 def test_workflow_step_factory_creates_step():
     step = WorkflowStepFactory()
     assert step.template.steps.count() == 1
+
+
+@pytest.mark.django_db
+def test_workflow_instance_factory_creates_instance():
+    instance = WorkflowInstanceFactory()
+    assert instance.project is not None
+    assert instance.status == instance.STATUS_DRAFT
